+++ lib/web_translate_it/project.rb
@@ -7,9 +7,7 @@ module WebTranslateIt
     def self.fetch_info(api_key) # rubocop:todo Metrics/MethodLength
       Util.with_retries do
         WebTranslateIt::Connection.new(api_key) do |conn|
-          request = Net::HTTP::Get.new("/api/projects/#{api_key}")
-          WebTranslateIt::Util.add_fields(request)
-          response = conn.http_connection.request(request)
+          response = conn.get("/api/projects/#{api_key}")
           return response.body if response.is_a?(Net::HTTPSuccess)
 
           puts 'An error occured while fetching the project information:'
@@ -26,27 +24,21 @@ module WebTranslateIt
       url = file_id.nil? ? "/api/projects/#{api_key}/stats" : "/api/projects/#{api_key}/stats?file=#{file_id}"
       Util.with_retries do
         WebTranslateIt::Connection.new(api_key) do |conn|
-          request = Net::HTTP::Get.new(url)
-          WebTranslateIt::Util.add_fields(request)
-          return Util.handle_response(conn.http_connection.request(request), true)
+          return Util.handle_response(conn.get(url), true)
         end
       end
     end
 
     def self.create_locale(connection, locale_code)
       Util.with_retries do
-        request = Net::HTTP::Post.new("/api/projects/#{connection.api_key}/locales")
-        WebTranslateIt::Util.add_fields(request)
-        request.set_form_data({'id' => locale_code}, ';')
-        Util.handle_response(connection.http_connection.request(request), true)
+        response = connection.post("/api/projects/#{connection.api_key}/locales") { |req| req.set_form_data({'id' => locale_code}, ';') }
+        Util.handle_response(response, true)
       end
     end
 
     def self.delete_locale(connection, locale_code)
       Util.with_retries do
-        request = Net::HTTP::Delete.new("/api/projects/#{connection.api_key}/locales/#{locale_code}")
-        WebTranslateIt::Util.add_fields(request)
-        Util.handle_response(connection.http_connection.request(request), true)
+        Util.handle_response(connection.delete("/api/projects/#{connection.api_key}/locales/#{locale_code}"), true)
       end
     end